requirements.txt
Django==3.0.5
#torch==1.5.0+cpu
#torchvision==0.6.0+cpu
numpy==1.18.2
opencv-python==4.2.0.34
tensorflow==2.1.0
Pillow==7.1.1
keras==2.3.1
h5py==2.10.0
pandas==1.0.3
